+++ examples/demo/margin-control.c
@@ -1,6 +1,6 @@
 /* This entire file is licensed under MIT
  *
- * Copyright 2020 William Wold
+ * Copyright 2020 Sophie Winter
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
  *
@@ -100,6 +100,7 @@ margin_control_new (GtkWindow *layer_window, const int default_margins[GTK_LAYER
     }
 
     GtkWidget *open_button = gtk_button_new_with_label ("Set margin");
+    gtk_widget_set_tooltip_text (open_button, "Space to leave empty around surface");
     GtkWidget *popover = gtk_popover_new (open_button);
     gtk_popover_set_modal (GTK_POPOVER (popover), TRUE);
     gtk_popover_set_constrain_to (GTK_POPOVER (popover), GTK_POPOVER_CONSTRAINT_WINDOW);