+++ gtk-priv/h/common.h
@@ -1,6 +1,6 @@
 /* This entire file is licensed under MIT
  *
- * Copyright 2020 William Wold
+ * Copyright 2020 Sophie Winter
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
  *
@@ -27,16 +27,18 @@ void gtk_priv_warn_gtk_version_may_be_unsupported() {
     static gboolean shown = FALSE;
     if (shown)
         return;
-    g_warning(
-        "gtk-layer-shell v%d.%d.%d may not work on GTK v%d.%d.%d. "
-        "If you experience crashes, check "
-        "https://github.com/wmww/gtk-layer-shell/blob/master/compatibility.md",
-        GTK_LAYER_SHELL_MAJOR,
-        GTK_LAYER_SHELL_MINOR,
-        GTK_LAYER_SHELL_MICRO,
-        gtk_get_major_version(),
-        gtk_get_minor_version(),
-        gtk_get_micro_version());
+    if (getenv("GTK_LAYER_SHELL_UNSUPPORTED_GTK_WARNING")) {
+        g_warning(
+            "gtk-layer-shell v%d.%d.%d may not work on GTK v%d.%d.%d. "
+            "If you experience crashes, check "
+            "https://github.com/wmww/gtk-layer-shell/blob/master/compatibility.md",
+            GTK_LAYER_SHELL_MAJOR,
+            GTK_LAYER_SHELL_MINOR,
+            GTK_LAYER_SHELL_MICRO,
+            gtk_get_major_version(),
+            gtk_get_minor_version(),
+            gtk_get_micro_version());
+    }
     shown = TRUE;
 }