+++ test/run-integration-test.py
@@ -2,7 +2,7 @@
 '''
 This entire file is licensed under MIT.
 
-Copyright 2020 William Wold
+Copyright 2020 Sophie Winter
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
 
@@ -25,12 +25,6 @@ from typing import List, Dict, Optional, Any
 
 # All callables (generally lambdas) appended to this list will be called at the end of the program
 cleanup_funcs = []
-
-class Color:
-    '''Unix terminal color escape sequences'''
-    normal = '\x1b[0m'
-    gray = '\x1b[90m'
-    decoration = gray
 
 class TestError(RuntimeError):
     pass
@@ -74,14 +68,11 @@ def format_stream(name: str, stream: str) -> str:
         r_pad -= 1
     l_pad = max(l_pad, 1)
     r_pad = max(r_pad, 1)
-    header = '─' * l_pad + '┤ ' + Color.normal + name + Color.decoration + ' ├' + '─' * r_pad + '┈'
-    divider = '\n' + Color.decoration + '│' + Color.normal
+    header = '─' * l_pad + '┤ ' + name + ' ├' + '─' * r_pad + '┈'
+    divider = '\n│'
     body = divider.join(' ' + line for line in stream.strip().splitlines())
     footer = '─' * 60 + '┈'
-    return (
-        Color.decoration + '╭' + header + Color.normal +
-        divider + body +
-        '\n' + Color.decoration + '╰' + footer + Color.normal)
+    return '╭' + header + divider + body + '\n╰' + footer
 
 def format_process_report(name: str, returncode: int, stdout: str, stderr: str) -> str:
     '''After running a program, this function is used to format it's output for easy reading'''